pyproject.toml
[tool.poetry]
name = "ragapp"
version = "0.1.0"
description = ""
authors = ["jaipaulmann <71453042+jaipaulmann@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
awscli = "^1.32.117"
boto3 = "1.34.123"
streamlit = "1.35.0"
pypdf2 = "^3.0.1"
langchain = "0.2.3"
openai = "^1.30.5"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
altair = "5.3.0"
annotated-types = "0.7.0"
async-timeout = "4.0.3"
attrs = "23.2.0"
blinker = "1.8.2"
botocore = "1.34.123"
cachetools = "5.3.3"
certifi = "2024.6.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
dataclasses-json = "0.6.7"
frozenlist = "1.4.1"
gitdb = "4.0.11"
gitpython = "3.1.43"
idna = "3.7"
importlib-resources = "6.4.0"
jinja2 = "3.1.4"
jmespath = "1.0.1"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
jsonschema = "4.22.0"
jsonschema-specifications = "2023.12.1"
langchain-aws = "0.1.6"
langchain-community = "0.2.4"
langchain-core = "0.2.5"
langchain-text-splitters = "0.2.1"
langsmith = "0.1.76"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
marshmallow = "3.21.3"
mdurl = "0.1.2"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
numpy = "1.24.4"
orjson = "3.10.4"
packaging = "23.2"
pandas = "2.0.3"
pillow = "10.3.0"
pkgutil-resolve-name = "1.3.10"
protobuf = "4.25.3"
pyarrow = "16.1.0"
pydantic = "2.7.3"
pydantic-core = "2.18.4"
pydeck = "0.9.1"
pygments = "2.18.0"
python-dateutil = "2.9.0.post0"
pytz = "2024.1"
pyyaml = "6.0.1"
referencing = "0.35.1"
requests = "2.32.3"
rich = "13.7.1"
rpds-py = "0.18.1"
s3transfer = "0.10.1"
six = "1.16.0"
smmap = "5.0.1"
sqlalchemy = "2.0.30"
tenacity = "8.3.0"
toml = "0.10.2"
toolz = "0.12.1"
tornado = "6.4.1"
typing-inspect = "0.9.0"
typing-extensions = "4.12.2"
tzdata = "2024.1"
urllib3 = "1.26.18"
watchdog = "4.0.1"
yarl = "1.9.4"
zipp = "3.19.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
